% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bioenergy}
\alias{bioenergy}
\title{FAOSTAT Bioenergy Database}
\format{
A list of time series.
}
\source{
\href{http://www.fao.org/faostat/en/#data/BE}{FAOSTAT Bioenergy Database}
}
\usage{
data(bioenergy)
}
\description{
Bioenergy data from FAOSTAT.
Data Type:	Bioenergy consumption and production. Category:	Environment. Creation Date	2024.
}
\examples{
data(bioenergy)
serie <- bioenergy[[1]]
ts.plot(serie)
}
\references{
FAO 2024. FAOSTAT Bioenergy, FAO, Rome, Italy.  ; United Nations Statistics Division (UNSD), 2011; International Recommendations for Energy Statistics (IRES).
}
\keyword{datasets}
