% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gecco}
\alias{gecco}
\title{Gecco Challenge 2018}
\format{
A list of time series.
}
\source{
\href{https://www.spotseven.de/gecco/gecco-challenge}{Gecco Challenge 2018}
}
\usage{
data(gecco)
}
\description{
Data collection for water quality monitoring.
Data Type:	Water quality. Category:	Environment. Creation Date	2018.
}
\examples{
data(gecco)
serie <- gecco[[1]]
ts.plot(serie)
}
\references{
Genetic and Evolutionary Computation Conference (Gecco), Association for Computing Machinery (ACM)
}
\keyword{datasets}
