% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m3}
\alias{m3}
\title{M3 Competition Time Series}
\format{
A list of lists containing time series.
}
\source{
\href{https://www.sciencedirect.com/science/article/abs/pii/S0169207000000571?via\%3Dihub}{The M3-Competition: results, conclusions and implications}
}
\usage{
data(m3)
}
\description{
Time series data from the third Makridakis forecasting competition (M3), held in 2000.
Data Type: Forecasting benchmark dataset. Category: Forecasting. Creation Date: 2000.
}
\examples{
data(m3)
serie <- m3$monthly$M1
ts.plot(serie)
}
\references{
Makridakis and Hibon (2000). The M3-Competition: Results, conclusions and implications. International Journal of Forecasting, 16(4), 451–476.
}
\keyword{datasets}
