% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{m4}
\alias{m4}
\title{M4 Competition Time Series}
\format{
A list of lists containing time series.
}
\source{
\href{https://github.com/Mcompetitions/M4-methods}{M4 Competition - GitHub}
}
\usage{
data(m4)
}
\description{
Time series data from the fourth Makridakis forecasting competition (M4), held in 2018.
Data Type: Forecasting benchmark dataset. Category: Forecasting. Creation Date: 2018.
}
\examples{
data(m4)
serie <- m4$monthly$M1
ts.plot(serie)
}
\references{
Makridakis et al. (2020). The M4 Competition: Results, findings, conclusion and way forward. International Journal of Forecasting, 36(1), 54–74.
}
\keyword{datasets}
