% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pesticides}
\alias{pesticides}
\title{Pesticides Use Statistics (1990–2022)}
\format{
A list of time series.
}
\source{
\href{http://www.fao.org/faostat/en/#data/RP}{http://www.fao.org/faostat/en/#data/RP}
}
\usage{
data(pesticides)
}
\description{
Statistics on the use of major pesticide groups and relevant chemical families.
Data Type: pesticides use. Category:	Environments. Creation Date	2024.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(pesticides)
serie <- pesticides[[1]]
}
\references{
FAO. 2024. FAOSTAT: Pesticides Use. RP_e_README_Domain_Information_2024.
\href{http://www.fao.org/faostat/en/#data/RP}{FAOSTAT Pesticides Use Database}
}
\keyword{datasets}
