% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SantaFe.A}
\alias{SantaFe.A}
\title{Santa Fe Time Series Competition - Series A}
\format{
A data frame with one column and 1,100 rows, containing numeric time series values.
}
\source{
\href{https://archive.physionet.org/physiobank/database/SantaFe.A/}{Santa Fe Time Series Competition}
}
\usage{
data(SantaFe.A)
}
\description{
Univariate time series A from the Santa Fe Time Series Competition.
Data Type: Laser-generated nonlinear time series. Category: Benchmark. Observations: 1,100.
This benchmark dataset consists of a low-dimensional nonlinear and stationary series recorded from a Far-Infrared-Laser in a chaotic regime.
Competitors were asked to predict the last 100 observations, and performance was evaluated using NMSE.
}
\examples{
data(SantaFe.A)
serie <- SantaFe.A$V1
ts.plot(serie)
}
\references{
Weigend, A.S. (1993). \emph{Time Series Prediction: Forecasting the Future and Understanding the Past}. Reading, MA: Westview Press.
}
\keyword{benchmark}
\keyword{datasets}
