% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stocks}
\alias{stocks}
\title{IBOVESPA's 50 Most Traded Stocks}
\format{
A list of time series.
}
\source{
\href{https://www.b3.com.br}{B3}
}
\usage{
data(stocks)
}
\description{
Historical daily data for the 50 most traded stocks in B3 (IBOVESPA),
including opening, high, low, and closing prices, as well as trading volume.
Data Type: Financial Time Series. Category: Finance. Creation Date: 2025.
}
\examples{
data(stocks)
serie <- stocks$VALE3$close
ts.plot(serie)
}
\references{
B3 - Brasil, Bolsa, Balcão. 2025. Historical stock trading data.
\href{https://www.b3.com.br}{B3 Official Website}
}
\keyword{datasets}
