% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fertilizers}
\alias{fertilizers}
\title{FAOSTAT Fertilizers by Nutrient}
\format{
A list of time series.
}
\source{
\href{https://www.fao.org/faostat/en/#data/RFN}{FAOSTAT Fertilizers by Nutrient}.
}
\usage{
data(fertilizers)
}
\description{
Statistics on agricultural use, production, and trade of chemical and mineral fertilizers.
Data Type:	Fertilizers use, production and trade. Category:	Environment. Creation Date	2024.
See \href{https://github.com/cefet-rj-dal/event_datasets}{cefet-rj-dal/event_datasets}
for detailed guidance on using this package and the other datasets available in it.
Labels available? Yes
}
\examples{
data(fertilizers)
serie <- fertilizers[[1]]
}
\references{
FAO, 2024. FAOSTAT: Fertilizers by Nutrient.

FAO & UNSD (2017). System of Environmental-Economic Accounting for Agriculture, Forestry and Fisheries (SEEA AFF).

UNSD (2017). Framework for the Development of Environment Statistics (FDES).
}
\keyword{datasets}
