% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pesticides}
\alias{pesticides}
\title{Pesticides Use Statistics}
\format{
A list of time series.
}
\source{
\href{http://www.fao.org/faostat/en/#data/RP}{Pesticides Use Database}
}
\usage{
data(pesticides)
}
\description{
Statistics on the use of major pesticide groups and relevant chemical families.
Data Type: pesticides use. Category:	Environments. Creation Date	2024.
}
\examples{
data(pesticides)
serie <- pesticides[[1]]
ts.plot(serie)
}
\references{
FAO. 2024. FAOSTAT: Pesticides Use. RP_e_README_Domain_Information_2024.
\href{http://www.fao.org/faostat/en/#data/RP}{FAOSTAT Pesticides Use Database}
}
\keyword{datasets}
